refactor: Let the stream session own what lasts one stream
Closes #20.

"Only while the main broadcaster is live" was answered three ways: a
boolean on the shoutout queue, the presence of a task in a dict on the
session, and a live_alert row plus a Helix call. CONTEXT.md already said
a session owns the queue and the ad-break warning; the code did not.

The session now holds the Stream it began for, and is the only answer to
whether the bot is live. It holds the Stream rather than a flag because
an offline check that took a moment has to tell "the stream I asked
about is gone" from "a different one began while I was asking" —
stream_online can sit in _wait_for_stream_info for a minute before
began() runs, so that race is wide, not theoretical.

stream.offline now wakes the session as well as the alert updater, and
each re-checks Helix for its own scope. live_alert no longer calls
ended(): that made the session's correctness depend on an alert row
existing, and three paths end an updater without reaching the call —
_decide returning STOP on a missing row, _run giving up after
_MAX_INCONCLUSIVE_CYCLES, and a failed announce() starting no updater at
all. In each, the stream ended and the queue went on believing it was
live. ADR 0004 records the decision.

The drainer is one task for the life of the process, started in
setup_hook rather than on_ready, which fires again on every gateway
reconnect. That retires the generation counter, which existed only
because a drainer told to stand down could still be inside a two-minute
sleep when the next stream started another alongside it. It still asks
the session each pass, and that is load-bearing: a target whose lookup
failed re-queues itself, so a lookup failing as the stream ended would
otherwise have been shouted out into an offline chat. The 60-minute
per-target cooldown deliberately survives a session, since Twitch keeps
it whatever this process does.

The ad-break dict collapses to one field. It was keyed by broadcaster
because the subscription is, but only the main broadcaster has that
subscription, so it never held two; scheduling is now guarded on the
main broadcaster instead, and the identity-checked done-callback and its
lambda capture are gone.

Three findings the gate raised on pre-existing code, fixed here rather
than waived because you asked for them in this PR:

- live_alert.py was over the 400-line limit. Split by responsibility:
  live_alert.py starts, wakes and restores updaters; live_alert_cycle.py
  is one pass — _decide and the refresh and close that carry its answer
  out — and knows nothing about tasks. 485 lines becomes 226 and 292.
- The assert in the REFRESH path is now an explicit guard. -O strips an
  assert, which would have left _refresh taking a None it is not typed
  for.
- !so validates its target against the Twitch login grammar before the
  lookup, so a word that cannot name a channel costs no Helix call and
  never reaches Twitch as a query parameter.

A fourth, "self-referential class annotation raises at import time", is
false on this interpreter: Python 3.14 evaluates annotations lazily
(PEP 649), the module imports cleanly and its __annotations__ resolve.
Reported as a false positive against both runs rather than waived.

Verified with the four checks in order, all clean, and a throwaway
harness covering the supersede race, the offline straggler, the cooldown
surviving, the ad-break slot and the login grammar. There is no test
suite, so this is not tested — it has not yet run against a real stream.
